Resets for a refactor to chain transformations

+++ rust/src/measurements/make_private_expr/expr_dp_frame_len/mod.rs
@@ -48,7 +48,7 @@ impl AnonymousColumnsUdf for DPFrameLenShim {
         _: &polars::prelude::Schema,
         _: &[polars::prelude::Field],
     ) -> PolarsResult<polars::prelude::Field> {
-        Ok(Field::new("len".into(), DataType::Int64))
+        Ok(Field::new("len".into(), DataType::UInt32))
     }
 }
 impl OpenDPPlugin for DPFrameLenShim {
